requirements.txt
certifi==2020.6.20
chardet==3.0.4
Django==2.2
django-cors-headers==3.4.0
djangorestframework==3.11.1
idna==2.10
mysqlclient==2.0.1
oauthlib==3.1.0
PySocks==1.7.1
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
tweepy==3.9.0
urllib3==1.25.10
